pyproject.toml
[project]
name = "3d-recon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.33.1",
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "pyyaml>=6.0.2",
    "rich>=13.7.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]
